refactor: rename ProcessID to mark it as a Raw value
