+++ rt/src/deserialize.mts
@@ -4,7 +4,7 @@ import { spawn, ChildProcess } from 'child_process'
 import * as Ty from './base/TroupeTypes.mjs'
 import { isLVal, LVal } from './base/LVal.mjs';
 import { mkTuple, mkList, mkRecord } from './base/rawUtil.mjs';
-import { ProcessID } from './base/ProcessID.mjs';
+import { RawProcessID } from './base/RawProcessID.mjs';
 import { RawAuthority } from './base/RawAuthority.mjs';
 import { Atom } from './base/Atom.mjs';
 import RawUnit from './base/RawUnit.mjs'
@@ -267,7 +267,7 @@ async function reconstruct(jsonObj: any, compilerOutput: string | undefined, tru
                 case Ty.TroupeType.STRING:
                     return obj;
                 case Ty.TroupeType.PROCESS_ID:
-                    return new ProcessID(obj.uuid, obj.pid, obj.node)
+                    return new RawProcessID(obj.uuid, obj.pid, obj.node)
                 case Ty.TroupeType.AUTHORITY:
                     // Attenuate authority based on the trust level of the sender
                     return new RawAuthority(_trustGLB(mkLevel(obj.authorityLevel)));
